+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Wang 2019 cyclosporin ([doi:10.3892/etm.2019.7325](https://doi.org/10.3892/etm.2019.7325)) -- Chinese children with refractory nephrotic syndrome.
+- Add Meister 2019 tribendimidine ([doi:10.1128/AAC.01391-18](https://doi.org/10.1128/AAC.01391-18)) -- Lao adolescents and adults infected with the liver fluke *Opisthorchis viverrini*.
 
 - Add Palmer 2025 moxifloxacin ([doi:10.1002/bcp.70005](https://doi.org/10.1002/bcp.70005)) -- children with rifampicin-resistant tuberculosis.